fix(tui): resize at startup
The deleted code is not needed after 402b4e8

It caused the condition to false positive when the function was called
more than once before startup (first normal mode). Being itself what set
the dimension and not the user, locking the size of the screen to an
incorrect size.

Make got_winch an int to mitigate: tui_grid_resize changing the size of
the host terminal between the signal handler and the call to
sigwinch_cb. Since the actual signal handler uv__signal_handle doesn't
directly call the callback, the event loop does.