fix(treesitter): ensure `tstree:parse()` return trees and changes
When parsing from an invalidated state, `tstree:parse()` returns the
trees and the changes.

But when calling `tstree:parse()` again, with a valid state, the
function returns only the trees.
This would change the behavior of calls like `tstree:parse()[1]`
expecting to get the list of trees, and instead getting the first tree.