feat(colorcolchar): add the option "colorcol" to the fillchars setting
This option will let neovim draw a character in the colorcolumn when
there is no other character occupying that spot.

For example, I'm someone who likes the elegance of seeing a 1px wide line
at the 80 character mark, rather than a rectangle the width of a cell at
that mark. To accomplish this, I run

:set colorcol=80
:set fillchars=colorcol:│

of course ':' and '.' are good ASCII alteratives.