build: only generate compilation database for the nvim target
Some tools like clang-tidy get confused and repeat unnecessary work if
there are multiple entries with the same name.

This will only generate a compilation database for cmake version 3.20
and above. Generating a compilation database is only necessary for
development, so we don't need to maintain compatibility with the minimum
required version.

Closes https://github.com/neovim/neovim/issues/10632