fix(build): close uv handles on a different define
Previously if EXITFREE was defined, then uv handles were not closed.
This was done to better detect memory leaks.

However, now EXITFREE is always defined for debug builds which makes
them pretty annoying to use since not closing the handles causes a lag
on exit.

This change adds a new define, NOT_UV_CLOSE, which when defined will not
close the uv handles when exiting. This is automatically set when
sanitizers are enabled.