build: remove EXITFREE for debug builds
When EXITFREE is defined, uv handles are not closed on exit in order to
better detect memory leaks. However the cost of this is that by not
closing the handles there can be lag when exiting nvim.

This change removes EXITFREE from debug builds in order to avoid this
exit lag for regular debug builders who are not running ASAN.