build(deps): specify C++ compiler for msgpack if LLVM/clang-cl is used
Problem:
The "msgpack" project has a dependency on C++ compiler.
If LLVM/clang-cl is used as the C compiler, it shall also be used as the C++ compiler for consistency.
However, the C++ compiler for "msgpack" is not configured in "BuildMsgpack.cmake".
And cmake would use LLVM/clang++ (the GNU frontend) as the C++ compiler, which is inconsistent with LLVM/clang-cl (the MSVC frontend).

Solution:
Specify C++ compiler for msgpack if LLVM/clang-cl is used as the C compiler.