fix(docs-html): fix tab alignment for concealed characters
The gen_help_html.lua script is a useful tool to generate structured
HTML pages from the vim docs, but it has an issue of not properly
handling tab characters under certain circumstances.

In Vim, sometimes the metadata is hidden using conceal, e.g.
*some_tag_definition*, |some_tag_link|, and `some_code_block`. For
convenience, Vim hides the '*' / '|' / '`' characters to make it more
readable, but those characters still count when a tab character follows
it. So if you have a tag of say 6 characters long, those two concealed
character would lead to the tab character after it start at character 8.
The gen_help_html doesn't account for that which is what leads to a lot
of formatting errors in the generated output.

This commit fixes that by manually adding two spaces after each
concealed tag, if it's followed by a tab char.