feat(autoread): added libuv watcher for autoread
Added a libuv fswatch watcher for autoread. The watcher starts
watching the current buffer when it is opened in neovim via
autocommands. If the watched buffer changes outside neovim, it
either automatically reloads the file or notifies the user about
the change depending upon the value of the option autoread.