feat(exrc): un-deprecate 'exrc' option
Add a function `read_secure` to read a file and ask a user if the file
should be trusted. The user's decision is remembered in a "trust
database" ($XDG_STATE_DIR/nvim/trust by default) which tracks the full
file path and the SHA256 hash of the file contents. This function is
used when the 'exrc' option is enabled to enable a more secure version
of the 'exrc' feature.