docs: upstream nanotee's Lua guide
nanotee's Lua guide (https://github.com/nanotee/nvim-lua-guide) has been
the semi-official Neovim-lua guide for a long time. Even the Neovim
documentation itself refers to it as the recommended source for learning
Lua in Neovim, and for good reason. There is no good reason not to have
this excellent material shipped with Neovim by default.

Important to note is that this isn't an exact port of the guide but a
heavily reworked version. Focus has been shifted from being a full,
comprehensive guide to a more "starter kit" to get someone up and
running fast. For example, if there are Nvim API and Lua API functions
to perform a given task, then the Lua API functions are presented.