feat(diagnostic): add `suffix` option to `virt_text` config
Follow-up to https://github.com/neovim/neovim/pull/21130

This introduces a `suffix` option to the `virt_text` config in
`vim.diagnostic.config()`. The suffix can either be a string which is appended
to the diagnostic message or a function returning such. The function receives a
`diagnostic` argument, which is the diagnostic table of the last diagnostic (the
one whose message is rendered as virt text).

We could also pass in `i` and `total` to the function, like we do for the
`suffix` in `vim.diagnostic.open_float()`. This would enable configuring a
suffix like " [3/3]" for a line with 3 diagnostics. I'm not sure how useful that
is, though.

Also unlike the `suffix` in `vim.diagnostic.open_float()`, we don't support
customizing the highlight. I'm not sure how useful this would be.