fix(syntax): Revert runtime/syntax/lua.vim to workaround broken syntax
There is a regression on the lua syntax file $VIMRUNTIME/syntax/lua.vim
in neovim 0.8.0, since commit or 67df3347f, where syntax is broken after
a lua function call (parenthesis). See #20456 for more details.

This is apparently because of a upstream vim syntax engine bug,
but at this moment we do not yet have an available fix. A reasonable
workaround until the upstream bug gets fixed would be to revert the
syntax file to d3068d3 (v0.7.2), which does not break the syntax
with the current vim syntax engine as of neovim 0.9.0-nightly and stable
versions (0.8.x).