feat(grep/make): read from stdout by default (remove shellpipe)
The 'shellpipe' option is not (that) necessary: read directly from stdout.
This simplifies some of the complexity introduced by an earlier commit
which special-cased an empty 'shellpipe' as meaning: no temporary files.
After this commit :make/:grep will never make temporary files themselves
anymore. The ability to source quickfix/loclist items from a file
remains: if 'makeef' is set, that will be used as the filename to read.