vim-patch:9.0.1100: a hashtab with many removed items is not cleaned up
Problem:    A hashtab with many removed items is not cleaned up.
Solution:   Re-hash a hashtab even when the size didn't change if too many
            items were removed.

https://github.com/vim/vim/commit/d0883faac6a74f777c9a6be9d035c59ee1c969c5

N/A patches for version.c:

vim-patch:9.0.1099: trying to resize a hashtab may cause a problem

Problem:    Trying to resize a hashtab may cause a problem.
Solution:   Do not try to resize a hashtab before adding an item.

https://github.com/vim/vim/commit/81b7ecc5cb78cad901a9a85e46ecba109cc6ee7d

Co-authored-by: Bram Moolenaar <Bram@vim.org>