build(nix): remove incompatible reproducible patch
NixOS/nixpkgs#208103 introduced a patch that aims to make builds reproducible.
This patch applies cleanly against 0.8.2, but doesn't apply against
master.

Since the neovim flake uses the "upstream" nixpkgs neovim package, this
patch is also pulled in when trying to build the nightly version.

This PR simply removes the reproducible patch from the list of patches
that is applied, thereby allowing the master branch to be built once
again.